Move embedded video on SQL Developer from ADW Lab 3 to Lab 7
Per Nilay, since Lab 3 focuses on SQL Developer Web, it makes no sense to have a video on SQL Developer in this lab. Removed the video from Lab 3.
Lab 7 had a NON-EMBEDDED link to this video, causing a new browser tab to open in YouTube. Changed this link to use show the video embedded in Lab 7, so viewer can watch without being transported out of the lab.