Update the master learning-library with BDS workshops new folder structure
- Added a "workshops" folder under the "\big-data\bds" folder on my clone.
- Added the "freetier" and "livelabs" folders under "workshops" to match the template structure.
- Copied all the files that I had in the "bds-quickstart-non-ha-workshop" workshop folder to the "\workshops\freetier" folder.
- Removed the bds-quickstart-non-ha-workshop.
- Edit the manifest.json file to reflect the new pathing (all labs but not the intro).
For example:
"filename": "./../bds-create-cluster/bds-create-cluster.md" changes to
"filename": "./../../bds-create-cluster/bds-create-cluster.md"
We need to go two levels up to access this .md file: from the current "freetier" folder, we need to go up to the "workshops" folder, and then one more level up to the "bds-create-cluster" folder to get to the bds-create-cluster.md file.

Repeated the same process for the bds (HA) workshop.