Temporarily remove auto scaling lab from my fork so I can create pull request for a different workshop
This auto scaling lab is still under review by ADW Product Management. I've had it on my personal fork to allow reviewers to see preview how it will appear, and to be able to copy/paste the lab into a Confluence page for reviewers to be able to leave review comments. The copy/paste from personal GitHub fork allows reviewers to see images. I need to create a pull request for a different workshop, so need to temporarily remove the auto scaling lab that's still in review, so it doesn't get pulled into production.