removal of auto scaling lab and ODI/ADW workshop from my local clone
push the removal of auto scaling lab and ODI/ADW workshop from my local clone, up to my fork, before making a pull request to remove them from production repo. After the pull is approved and merged, I'll add the auto scaling lab back to clone,  push it to fork, examine it on the fork, and request to pull it into production repo.