Need Help update for: ALL APEX workshops
There seems to be a few old or defunct workshops that have a different banner and page layout (Usually when the manifest does NOT include a "workshoptitle" attribute). I've added the "help" attribute to those manifest.json's but they... don't manifest (teehee)... on the page, the additional ? button doesn't populate.  Pretty sure it'd be a fast fix, add the new script link to each index.html