Interim update of Autonomous Database Quickstart workshop
This interim update of the workshop is limited to changing the use of SQL Developer Web to its successor, SQL Worksheets. There are changes to procedural steps and screenshots. Later this week or early next week I'll finish the major update, which adds Database Actions to the "Loading Data" lab for easier data loading.