Change one word in Enabling Data Integrity lab to try wake up github.io to refresh
Since Thursday, github.io pages haven't refreshed to show changes, even though the pushed .md file does show changes in my personal fork. This is an attempt to use our old trick of "waking up" github.io to refresh to show changes, when it fails to do so the first time. This is my second attempt with this trick.