Workshop ID#6481 fix Lab 3 code snippet to remove reference to db_user variable
We switched last week from the buggy UI to SQL Worksheet to grant additional roles. We first tried a variable for db_user, but when we got errors, we switched back to simply specifying "moviestream" as the user name. But we missed fixing the line:  p_schema => '&db_user'
I changed it to  p_schema => 'moviestream', so the learner isn't prompted for a db_user substitution variable.