Workshop (ID# 7041) Modified as per Self QA and lint checker along with other changes
Hello,

I performed Self QA and ran through the lint checker on the workshop.

Following are the modifications -

Removed the gerund and modified the title to Install Oracle Database 21c on OCI Compute.

Added these to the Introduction:
  - Estimated workshop time = 2 hours
  - Note on always free accounts. Used a new file to include this note.

Changed the folder names from 'common-files' to 'files' and from 'common-images' to 'images'.

Modified the manifest files as per the above changes.

Added the correct version of 'Need Help' in the manifests for LiveLabs and Main.

Removed the redundant file on 'See an issue'.

Minor changes to the lab Create Container Database (Advanced Mode)

Request for approval please.  Thank you
~MG