Workshop ID#141 modify script as workaround to software bug
There is a recently discovered bug in SQL Worksheet, in which Lab 4, Task 1, step 9 script fails to create external tables because there are more than 2 begin/end statements within the script. The workaround was to use just a single begin and end for the creation of all 9 external tables. We'll keep this workaround in place until Development fixes the bug.