Workshop ID#6481 Updated labs for new Database Actions button and Database Actions Launchpad
This is Marty's longer version, "Action! MovieStream Analytics Produce the Best Picture with Oracle Cloud - The Epic". This is NOT the "Shorts" version used in Oracle Database World workshop. I will submit the updates for the "Shorts" version as a separate pull request.
Here most of the updates involve the new Database Actions button and the updated Database Actions Launchpad page.