ADB-DCAT Workshop: Updated freetier and new LiveLabs versions
This is the updated version of the Access the Data Lake using Autonomous Database and Data Catalog workshop. The freetier will replace the version that I have in Production. The LIveLabs version is new and will have to coordinate with the LL team to make it available on LL; i.e. green button.